[Raspi] Fix EGL context initialization failure (#6627)
The evergreen qa build on Raspberry Pi is crashing due to a GL context
initialization failure. This is caused by the addition of EGL 1.5
attributes that are not supported on Raspi2.

This change fixes the crash by checking the EGL version and only adding
the attributes if the version is greater than 1.4.

Bug: 433739466

original-hexsha: 3640860ab39b0334836307ed4b1dcafb96a04b20