fix: Prevent white screen flash on startup (#6663)
On application launch, a brief white flash was visible before the first
frame was rendered. This was caused by the default-initialized window
surface being displayed before Cobalt had drawn its initial content.

With this fix, the app run into a brief grey flash. It is caused by
Kabuki.
If the app start the page from https://blackscreen.app/, there is not 
grey flash.

Bug: 414415206
Change-Id: Ia29926dbf307e8b6030e4e4e678e6b6d791dfbc7