Add a simpler SBDStarboardApplication implementation to ApplicationDarwin
The tvOS part
-------------

Tests were previously being run with SbRunStarboardMain() and
SBDUIKitApplicationMain(). This starts the whole starboard::Application main
loop and causes the tests to never stop running on tvOS, as it is not a
Starboard platform and we do not integrate with the Starboard main loop
infrastructure at all.

ApplicationDarwin essentially needs to be made more similar to
ApplicationAndroid, which also does not need to integrate with the Starboard
event loop. Instead of dealing with PlaybackCapabilities and AudioSink in
Initialize() and Teardown(), which are not called in our case, use the
constructor and destructor instead. SetCommandLine() is now called manually
in the constructor as well.

A new inner Objective-C++ interface owned by ApplicationDarwin is now
responsible for initializing the main objects needed by the tvOS
code (window manager, DRM manager, EGL adapter etc).

All these changes are integrated into the Starboard testing infrastructure
by creating a ::testing::Environment subclass whose purpose is to create and
destroy ApplicationDarwin instances between each test. It is called by
test_main.cc.

This is enough to get a lot of NPLB tests to work fine without further
integration with Starboard, the only downside is that the Starboard
test infrastructure on tvOS now depends on `//base` to avoid having to
reimplement all the iOS integration done in `base/test`.

With these changes, it is now possible to remove uikit_application.mm
altogether.

Patch by @abhijeetk and myself.

The general changes
-------------------

Supporting this means touching some platform-independent bits as well: since
`test_main.cc` now depends on tvOS's `test_support` when targeting the
platform, it makes sense to move it to a separate source set to avoid
duplicating the same checks and dependencies in different locations.

Bug: 456692793
Bug: 433180638