starboard: Add tests for Starboard Extension for Features
This PR adds testing for the Starboard Extension that adds Chromium
Features to the Starboard level.

An extension_test is added to check that the extension api itself is
retrievable. We also add `features_extension_test.cc`, which tests the
usage of hte starboard::features::FeatureList class. This class is what
intializes and stores features and parameters that are used in the
Starboard level.

Currently, these tests do not run on CI as their test build target,
starboard_platform_tests, is not built.

Bug: 435467391