cobalt/shell: Remove origin trials deps (#6699)
Origin Trials are for experimenting with Web features that are not yet
ready for production. Cobalt has no need for such a feature, since it is
not delivering new features for apps, but delivering a base for browsing
and media playback.

Test: out/android-arm_devel/bin/run_cobalt_browsertests -v
Test: out/android-arm_devel/cobalt_unittests -v
Fixed: 431035934

original-hexsha: a365fae16e2eb07b1b39baf000e77b8bc06c46b5