starboard/nplb: FakeGraphicsContextProvider doesn't need an SbWindow (#6686)
This CL/PR removes the SbWindow creation logic from
FakeGraphicsContextProvider, and also the provisions
in NPLB build files to support SbWindow creation on
Android TV builds.

FakeGCP should not rely on SbWindow creation IIUC,
unless -guess- it's to transitively initialize the windowing
system in ways needed for graphics to operate and
that'd look like relying on an implementation detail which
should be removed anyway.

I followed the few instances I could find of the created
window being used and they ended up in nothing, albeit
partners might be using it. NOTE: @xiaomings pointed
out in a review comment that this might have been a
legacy of previous platforms not present in Chrobalt.

This PR/CL is a followup to #6494.

Bug: 297808555

original-hexsha: c3a63553ca18451450906de05d3dcd0c28dc9832