cobalt/shell: Remove services test_support
This change removes services/device/public/cpp:test_support, since it is unnecessary for production behavior.
This probably breaks some tests in content_browsertests, but that has yet to be fully ported to Cobalt and any relevant test code should be refactored into a test-specific target. It should not be in the release binary.

Test: out/android-x86_devel/bin/run_cobalt_browsertests -v
Test: out/android-x86_devel/cobalt_unittests -v
Fixed: 436872870