refactor: Make starboard::Queue use std::mutex (#6724)
This PR refactored `starboard::Queue` to use `std::mutex`, instead of
deprecated `starboard::Mutex`.

This PR also fixes a race condition in the destructor. The refactoring
of the class exposed a pre-existing race condition in the test helper.
The destructor was not safely ensuring that the EGL context was
destroyed on the correct thread before proceeding with teardown. This
has been corrected by using to guarantee synchronous execution and
prevent EGL errors during testing.

Bug: 390503926