cobalt/shell: Remove services test_support (#6739)
This change removes services/device/public/cpp:test_support, since it is
unnecessary for production behavior. This probably breaks some tests in
content_browsertests that will eventually be ported to Cobalt, but those
have yet to be fully ported to Cobalt and any relevant test code should
be refactored into a test-specific target. It should not be in the
release binary.

Note: location_system_permission_status.h is the include that makes the
`services/device/public/cpp` dependency necessary (for content shell).
Removing this, builds work, tests pass, and user journeys seem to be
working. So, it might just be a leftover artifact that is more difficult
to catch than unused Java imports.

Test: out/android-x86_devel/bin/run_cobalt_browsertests -v
Test: out/android-x86_devel/cobalt_unittests -v
Fixed: 436872870