Edit musl files to fully support utimes/futimes
While we have added a wrapper to utimensat, we found out that utimes and
futimes were not calling this wrapper. To ensure that this happens, we
edit some musl files to ensure that utimes/futimes can successfully find
utimensat.

The edits range from adding additional build files to the musl build.gn,
and adding/disabling code through the #if defined(STARBOARD) check.

Bug: 432315159