build: Refactor build files to use is_cobalt_hermetic_build (#6615)
This change replaces the `is_cobalt` and `!is_android` flags adjusting
the libc and ICU builds with the `is_cobalt_hermetic_build` flag.

The `is_cobalt_hermetic_build` flag is used to indicate that the build
should not have any dependencies on the host system's libraries, and
instead use the versions of libraries that are checked into the
repository.

This change affects the build configuration for libc and ICU.

Bug: 434866723

original-hexsha: 3e316ee8377eae974285a370ef46eb1479c62b23