Evergreen: replace tts_linux to tts_starboard stub
This PR adds tts_starboard.cc, which stubs in the
TtsPlatformImplBackgroundWorker and TtsPlatformImplLinux classes from
tts_linux.cc.

This allows us to remove our hermetic hack in tts_platform_impl.cc,
which was causing crashes on our evergreen builds. This hack
originally just returned nullpointer when the instance of the tts
implementation was called, which caused issues most likely due to code
assuming that this value isn't a nullptr.

Future PRs will properly implement this class.

Bug: 420913744
Bug: 437117010