media: Add max video size calculation experiment
This PR aims to re-add the code from pr/5565. which was closed due to
needing to be guarded by an experiment. With the ability to place
feature experiments inside Starboard now possible, we guard the code
from pr/5565 with a Starboard experiment.

When the starboard feature is enabled, we will calculate the max video
input size according to the latest exoplayer code from cl/467641494.
When the feature is disabled, we will use our current method of
calculating the max video input size.

Bug: 413696596