Change location of test file
While the feature extension is only accessible to the android platform
right now, it is possible in the future that different platforms would
want to make use of this extension. To account for this, we've moved the
location of the feature_extension test to starboard/shared/starboard,
allowing easy access for future platforms.

We've also changed the name of this file to feature_list_test.cc, as
these tests are designed to check the functionality of the
starboard/shared/starboard/feature_list implementation.

Bug: 435467391