fix(media): Use correct lock in MediaCodec callbacks (#6786)
This PR is the potential fix for the crash reported from Prime in
437914399

The problem is `MediaCodecBridge::stop()` and `onError` method
synchronize with different lock object, which can lead to the race
condition during destruction.
This PR made them use the same lock object.

per http://go/framework-code-patterns#heading=h.q8n720auvc61, this PR
changed the code to dedicate loc, in favor of this
#vibe-coded

Bug: 437914399