net: Restrict Cobalt trust store code to Linux (#6759)
The changes in #6602 broke the tvOS build for a few reasons:

* `test_root_certs_builtin.cc` is also added to the build in an `if
(is_mac)` block, which causes GN to complain.
* `SystemTrustStoreChromeOnly::chrome_root_store_version()` does not
exist on tvOS because `CHROME_ROOT_STORE_SUPPORTED` is not set.

Since this code seems to be trying to solve a Linux-specific problem,
restrict the changes to `net/BUILD.gn` to `is_linux`.

Bug: 432077378

original-hexsha: df9f587c25da9ecade719d45f3970696efe93049