HeapObserverSet to use a copy for iteration (#6815)
ContextLifecycleNotifier breaks the assumption that HeapObserverSet can
not be modified during iteration.
Before we identify the proper fix, let us try to iterate with a copy of
the observers and relax the write restriction.
Note that observer deletion during iteration might still cause crash but
from the top crash we observer it was AddObserver that's causing the
trouble.

Bug: 427982512