Enable basic crashpad functionality for evergreen
We're still considering some design decisions but for now this
implementation follows the general approach from c25: the handler
process is only launched after a crash and all interaction with the
crashpad client is done from below Starboard.

This is an initial step toward enabling crashpad: it may not yet be
portable and some major features, like runtime annotations and the use
of Cobalt's packaged CA root certificates, are still missing. But
crashes are captured and uploaded and minidumps can be symbolized for
crashes in the cobalt module.

Issue: 406511608