BACKPORT: build: Do not build mach module for macOS with Xcode earlier than 16.4
This backport contains only the //build part of the upstream commit, and is
required for us to be able to use `xcode_version_int` in GN coming from the
Mac side too.

Note that this backport does not actually use the new variable
anywhere (M138's `buildtools/third_party/libc++/BUILD.gn` does not contain
the parts being changed by this CL).

Bug: 562379920

---

Original CL description:

This is similar CL with https://crrev.com/c/6856959, but check Xcode
version for macOS build too.

This is to avoid error like
https://ci.chromium.org/ui/p/chromium/builders/try/ios-simulator/2445551/overview
which uses macOS SDK 15.2 which comes with Xcode 16.2.

Change-Id: I47a4a2e1f6ece213f32f3a5576dfaaa4d4212192
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6860827
Cr-Commit-Position: refs/heads/main@{#1503339}