android: Use Starboard features for use_stub_{audio/video}_decoder (#6890)
As `ApplicationAndroid` is created on the main thread, but
`PlayerComponents::Factory::CreateComponents()` is called from the media
thread, it is possible `shared::starboard::Application::Get()` is not
yet created when the media thread calls the function, resulting in a
native crash.

This PR moves `use_stub_{audio/video}_decoder` to use Starboard feature
on android, which may fix the crash.

Issue: 415049224