Check that pthread_create() succeeded, not the thread ID, in common code
`pthread_t` is an opaque type. It is an `unsigned long` in glibc and musl
when the latter is being used in C++ code, but it is an opaque pointer type
in musl in C code as well as in Apple's implementation.

This means that the recent changes from #6869 introduced some failures in
the (not-yet-upstreamed) tvOS build, as some SB_DCHECK_NE comparisons ended
up comparing a `pthread_t` (a pointer) to 0 (an integer).

Instead of comparing whether `pthread_create()` has changed the thread ID
pointer passed to it, it is more portable to just check whether the call to
`pthread_create()` itself has succeeded.

This commit does this for cross-platform code; there are other occurrences
in Android and Linux-specific code where this is less of a concern and which
would require changing other assertions that also compare `pthread_t`'s to
integers.

Fixed: 440370504