starboard: Add tests for Starboard Extension for Features (#6688)
This PR adds testing for the Starboard Extension that adds Chromium
Features to the Starboard level.

An extension_test is added to check that the extension api itself is
retrievable. We also add `feature_list_test.cc`, which tests the usage
of the starboard::features::FeatureList class. This class is what
intializes and stores features and parameters that are used in the
Starboard level. The parameter test was created by Gemini.

Currently, these tests do not run on CI as their test build target,
starboard_platform_tests, is not built. However, these tests were
checked locally and are verified that they passed locally.

Bug: 435467391