Initialize ICU correctly when ICU is not bundled with the binary (#6957)
Issue: 437151583

When the ICU file is bundled separately, ensure ICU initialization works
correctly
This code was initially introduced in #6048 and removed in #6817 . 
Re-adding it here after discussion with team members. We might have some
use cases in the future
where we need ICU bundled separately and hence would like to keep the
complex initialization alive for that case.