Add hermetic localtime, localtime_r, gmtime, gmtime_r, mktime, and timegm (#6748)
This adds `localtime`, `localtime_r`, `gmtime`, `gmtime_r`, `mktime`,
and `timegm` implementations that use ICU. These are functions that
convert between `time_t` and `struct tm` including applying the correct
timezone rules, short name and time offset lookups for `localtime`,
`localtime_r`, and `mktime`.

 It also adds extensive tests.

After this eztime can be deprecated. In order to not futher increase the
size of this PR, that will be done in a followup PR.

Bug: 406083355