Remove unused egl_swap_buffers.cc and related code (#6984)
This file is no longer used, so this change removes it and all related
code, including the `starboard_egl_wrapped_functions` static library and
the `SwapBuffersBegin` and `SwapBuffersEnd` functions. In Cobalt 25,
it's used in cobalt rendering backend, and that is gone entirely in
Cobalt 26.

Bug: 401006482

original-hexsha: 68cb6e256cd4ca7a68ad5aa2d07434110692ce0a