android/apk: Platform Error suspend on back button (#6889)
Currently, when the network error dialog is shown, the back button
reloads the application. The behavior should instead be to suspend the
app. This PR changes the behavior to suspend the app when the back
button is pressed on the dialog. Added a boolean flag to
CobaltActivity.java to reload the app onResume() in certain cases, as
the app would sometimes become unusable (gray screen) if the reload only
happened when dismissing the dialog in onDismiss().

Bug: 438634236
(cherry picked from commit f8db08e944a4d0edef8dd1b4bf33ad62a60a81a1)