starboard: Replace JniEnvExt with Jni helper methods (#7000)
This change refactors the JNI helper layer by replacing the `JniEnvExt`
class with the collection of helper methods.

The previous `JniEnvExt` inherited from `JNIEnv`, which forced callers
to perform a dangerous downcast from a `JNIEnv*` to a `JniEnvExt*` to
use its helper methods. This pattern was unsafe and also hid the
implicit management of JNI thread attachment within its `Get()` method.

This change replaces the class with a collection of standalone helper
functions. These functions have no internal state and explicitly require
a JNIEnv* as an argument. Callers are now responsible for obtaining the
JNIEnv*, which makes the code more transparent and explicit.

This refactoring achieves the following:
- Eliminates the dangerous downcast.
- Promotes clearer, safer, and more explicit JNI code.
- Makes JNI helper functions stand-alone methods from class member
methods.
-
Bug: 441310156