Add initial RDK build integration into cobalt 26 (#6854)
This adds RDK initial integration into cobalt 26. It is based on the
previous RDK support included in cobalt 25.

Please refer to commit msgs for more details about the changes.

Depends on
https://cobalt-review.googlesource.com/c/external/components/generic/cobalt/+/6910
(once this lands it has to be added either as a glient dep or as a
subtree under starboard/contrib/rdk - as done in 25.lts.1+).

Bug: 435468447