Fix race condition in TimeZoneState initialization
A pure virtual function call to `icu::TimeZone::getOffset` would occur when multiple threads tried to access timezone information simultaneously. This lead to a race condition where the `icu::TimeZone` object was not fully constructed before being used.

This was resolved by adding a mutex to `TimeZoneState` to ensure that the creation and access of the `icu::TimeZone` object is thread-safe. The `GetTimeZone()` method has been replaced with `WithTimeZone()` which executes a function within the mutex's protection.