starboard: Use different pthread_setname_np() on Apple platforms (#7052)
`pthread_setname_np()` is not a portable function (that's what the "np"
part stands for), and its signature differs between glibc/musl and
Apple's libc (the latter takes 1 argument, the former takes 2).

Chromium abstracts this via base's PlatformThread class, but Starboard
makes direct `pthread_setname_np()` calls that always assume the
glibc/musl signature is being used.

Since in bug 432507888 it was mentioned that it is preferred not to
ressurect `SbThreadSetName()`, add some `ifdef`s in code used by both
Linux-based platforms and tvOS so that the right function signature is
used depending on the build target.

Note that there are other occurrences in `//starboard`, but they are
Linux-specific and have not been updated.

Bug: 432507888

original-hexsha: 6b6289a8f925fb3d5bfcc58b9e3770258cf2bd25