Restore focus on window on resume (#7039)
Request focus on the root view. This is a simple and reliable way to
restore navigation control. The system will then find the first
focusable view within this root view to give it focus.

Test: Have problem testing deeplink on Cobalt, I think the deeplink is
configured to only Kimono. Also the bug is not reproducible anymore
maybe due to my Shield OS upgrade. But I believe this is a solid fix,
because I logged to see OnResume is called when app is resumed, and
calling the rootView.requestFocus(); function works fine.
I will follow up with QA to verify this fix.

Bug: 431281341