BACKPORT: mobile_config.gni: Stop unsetting target_platform on host toolchains
This is a backport of https://crrev.com/c/6317669 adapted to M114 and the
simplified version of the change that introduced the `target_platform`
variable.

---

Original change description:

Remove the part of https://crrev.com/c/6306066 that set
`target_platform` to an empty string when the host toolchain was being used.

The idea of preventing people from writing wrong GN code is good, but
there are not a lot of places that need to check this variable at the
moment and, most importantly, checking this value on the host toolchain
is the only way to know that this is a build targeting tvOS. This is
exactly what we need to do in V8's gni/v8.gni so that we always set
`v8_enable_lite_mode` on all toolchains.

Change-Id: Id8ac8683d288368d79b5013d05e6cb99c36b3b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6317669
Cr-Commit-Position: refs/heads/main@{#1427698}
(cherry picked from commit 901d870cbdf914f45dad72c71be52ee4e4a7a187)