BACKPORT: [tvOS] Add a separate BuildInfo.plist for tvOS builds
This is a backport of https://crrev.com/c/6450355 adapted to M114.

---

Original commit description:

It is identical to the iOS one, except the values in the UIDeviceFamily
array: the tvOS plist just has an ID for tvOS.

Not doing so causes the generated bundles (e.g. content_shell) to fail
to run:

```
App installation failed: Unable to Install “content_shell”
This app is not made for this device.
This app was not built to support this device family; app is compatible with (
    1,
    2
) but this device supports (
    3
)
Underlying error (domain=IXUserPresentableErrorDomain, code=6):
        Unable to Install “content_shell”
        This app is not made for this device.
```

Change-Id: I997d6d895a9472213afaa4571d12c4a06e283242
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6450355
Cr-Commit-Position: refs/heads/main@{#1447142}
(cherry picked from commit c2e062f1efdadc2570c7cef8e1330e55d2bce23d)