Add tvos-arm64-simulator Cobalt build config
This makes it possible to run e.g.

    cobalt/build/gn.py -p tvos-arm64-simulator -c debug

and generate `out/tvos-arm64-simulator_debug/` and build the `starboard`
target.

Note: as in the upstream build, `.gclient` needs `target_os=['ios']` so that
`gclient sync` fetches all required iOS/tvOS dependencies.

The focus of this commit is on adding all the required scaffolding: setting
the appropriate GN variables to the right values, adding tvOS configurations
to `cobalt/build` and `starboard/build`, adding the minimum amount of
C++ and GN configuration files under `starboard/tvos` and building a small
number of files under the `starboard` target.

The dependency on the cobalt targets in `//BUILD.gn` has been temporarily
disabled for tvOS builds because they depend on further changes to build
cobalt shell as a bundle, which will come later.

Bug: 432823234
Bug: 442923940