cobalt: Reenable UserAgentPlatformInfo unit tests (#7104)
This CL reenables the UserAgentPlatformInfo unit tests, which are
disabled on ToT. This would be a relatively simple operation except for
the fact that there're two calls into Starboad (namely
SbSystemGetProperty() and SbSystemGetExtension()) which make the unit
test not really a unit test (and also crashes on AndroidTV due to JNI
StarboardBridge not being properly loaded). This CL/PR just adds a
`...for_testing` flag to avoid those calls.

This CL/PR also makes InitializeUserAgentPlatformInfoFields() a member
function, which reduces verbosity and allows for the use of the
mentioned `...for_testing` flag.

This CL contains the less controversial bits of #7076

Bug: 442760345

original-hexsha: 64153cbc2241253dc8a6091005967af32589de24