Fix: check pthread_join return value
pthread_join is one of fundamental invariants in the code, ensuring that a thread is no longer running after it returns. Especially, no more callbacks are called from this point and after.  A failure in pthread_join breaches this invariant and leads to an state that code doesn't expect (e.g. the supposedly-completed thread is still calling the callbacks)