Fix: Regenerate iOS bundle file list (#6904)
This is a fix for chromium pre-commit checks to match the same
pre-commit checks Chromium would have run in an effort to better align
our code to Chromium's. You are being asked to review because you were
the last person to touch this file(s). If you think there's someone
better to review please add them. Please the review the changes and if
they look good please approve the PR.

Precommit error message:

Filelist needs to be re-generated. Please run 'python3
/usr/local/google/home/ericquinzhang/Documents/Cobalt/cobalt/chromium/src/build/ios/update_bundle_filelist.py
/usr/local/google/home/ericquinzhang/Documents/Cobalt/cobalt/chromium/src/media/unit_tests_bundle_data.filelist
/usr/local/google/home/ericquinzhang/Documents/Cobalt/cobalt/chromium/src/media/unit_tests_bundle_data.globlist
/usr/local/google/home/ericquinzhang/Documents/Cobalt/cobalt/chromium/src/'
and include the changes in this CL

Bug: 435503470