cobalt/build: Clear |media_use_ffmpeg| GN flag for all platforms (#7145)
`media_use_ffmpeg` is set to false for both Linux and Android. IIUC
it should be false for all platforms. This CL moves that configuration
to the upper level `configs/common.gn` so every current and future
platform (e.g. tvOS) sees the same setting.

Since `media_use_ffmpeg` is set to false, `ffmpeg_branding` does not
apply so this CL/PR removes it.

Bug: 380293712