Map the SbKey values to DOM keyboard events.
The SbKey is mapped to the 3 representation taken by the
ui::KeyEvent:
   1) SbKey to DomKey
   2) SbKey to DomCode
   3) SbKey to KeyboardCode

The fist two represent the current Web starndard, while the 3rd is
a legacy representation but still in use.

The key modifiers are also propagated from the SbInputData.
A simple unit test was added to validate that all keys that are covered
and that the undefined mappings are explicit.

Issue: 371429755